requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.2
bitarray==2.7.3
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
colabcode==0.3.0
comm==0.1.2
constantly==15.1.0
cryptography==39.0.2
cssselect==1.2.0
cytoolz==0.12.1
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.1
frozenlist==1.3.3
fsspec==2023.3.0
h11==0.14.0
hexbytes==0.3.0
huggingface-hub==0.13.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
ipfshttpclient==0.8.0a2
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter-server==1.23.6
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
lru-dict==1.1.8
lxml==4.9.2
magic-admin==0.0.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.3.7
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.4.3
netaddr==0.8.0
networkx==3.0
nltk==3.8.1
notebook==6.4.5
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.7.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
praw==7.7.0
prawcore==2.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.19.5
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyDispatcher==2.0.7
Pygments==2.14.0
PyJWT==2.6.0
pyngrok==5.2.1
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
queuelib==1.6.2
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
responses==0.18.0
rlp==2.0.1
Scrapy==2.8.0
scrapy-splash==0.9.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.13.2
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
tweepy==4.13.0
Twisted==22.10.0
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
uvicorn==0.13.1
varint==1.0.2
w3lib==2.1.1
wcwidth==0.2.6
web3==5.31.4
webencodings==0.5.1
websocket-client==1.5.1
websockets==9.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
zope.interface==5.5.2
